move krz specific envelope control processing into kuzweil specific code

+++ ork.lev2/inc/ork/lev2/aud/singularity/cz101.h
@@ -12,7 +12,7 @@ struct CzEnvelope {
   int _sustPoint = -1;
   bool _decreasing[8];
   float _time[8];
-  int _level[8];
+  float _level[8];
 };
 
 ///////////////////////////////////////////////////////////////////////////////
